chore: chaning example file location
