+++ example/RunRingBuffer.cpp
@@ -5,41 +5,34 @@
 
 template<typename T>
 rosneuro::DynamicMatrix<T> generate_frame(int nrows, int ncols, int index) {
-
 	rosneuro::DynamicMatrix<T> frame = rosneuro::DynamicMatrix<T>::Zero(nrows, ncols);
-
 	for (auto i = 0; i<frame.rows(); i++)
 		frame.row(i) = (index*nrows + i + 1)*rosneuro::DynamicMatrix<T>::Constant(nrows, ncols, 1);
 
 	return frame;
 }
 
-
 int main(int argc, char** argv) {
-	
 	ros::init(argc, argv, "test_ringbuffer");
 	
 	constexpr unsigned int frame_rows = 3;
 	constexpr unsigned int frame_cols = 10;
 
 	rosneuro::Buffer<float>* buffer = new rosneuro::RingBuffer<float>();
-	if(buffer->configure("RingBufferCfgTest") == false) {
+	if(!buffer->configure("RingBufferCfgTest")) {
 		ROS_ERROR("[%s] Configuration failed", buffer->name().c_str());
 		return false;
 	}
 	ROS_INFO("[%s] Configuration succeeded", buffer->name().c_str());
 
-
-	if(buffer->isfull() == false) 
+	if(!buffer->isfull())
 		ROS_INFO("[%s] Buffer is not full", buffer->name().c_str());
 
 	int niter = 0;
 	ROS_INFO("[%s] Filling the buffer", buffer->name().c_str());
-	rosneuro::DynamicMatrix<float> frame;
-	rosneuro::DynamicMatrix<float> output;
+	rosneuro::DynamicMatrix<float> frame, output;
 
 	try {
-
 		do {
 			std::cout<<"<<<<<<<<"<<std::endl;
 			frame = generate_frame<float>(frame_rows, frame_cols, niter);
@@ -48,7 +41,7 @@ int main(int argc, char** argv) {
 			output = buffer->get();
 			std::cout<<output<<std::endl;
 			niter++;
-		} while (buffer->isfull() == false);
+		} while (!buffer->isfull());
 	} catch (std::runtime_error& e) {
 		ROS_ERROR("%s", e.what());
 	}
@@ -56,7 +49,6 @@ int main(int argc, char** argv) {
 	ROS_INFO("[%s] Buffer filled after %d iterations", buffer->name().c_str(), niter);
 	output = buffer->get();
 	std::cout<<output<<std::endl;
-
 
 	ROS_INFO("[%s] Adding a new frame", buffer->name().c_str());
 	frame = generate_frame<float>(frame_rows, frame_cols, niter);
@@ -68,7 +60,6 @@ int main(int argc, char** argv) {
 
 	delete buffer;
 	return 0;
-
 }
